refactor(core): extract host-agnostic config parsers
