+++ packages/remnic-core/src/config.ts
@@ -3,8 +3,6 @@ import type {
   CodexCompactionFlushMode,
   CodingModeConfig,
   CodexCompatConfig,
-  ContradictionLocalizationConfig,
-  ContradictionScanConfig,
   DreamingConfig,
   DreamsLightSleepConfig,
   DreamsRemConfig,
@@ -43,6 +41,11 @@ import { parseRecallConcurrencyConfig } from "./recall-concurrency-config.js";
 import { parseExtractionLivenessConfig } from "./extraction-liveness.js";
 import { parseReplicaPeersConfig } from "./replica-peers-config.js";
 import { parseDependencyPropagationConfig } from "./dependency-propagation-config.js";
+import {
+  parseContradictionLocalizationConfig,
+  parseContradictionScanConfig,
+} from "./contradiction-config.js";
+import { parseGraphPathScoringConfig } from "./graph-path-scoring-config.js";
 import { hasLegacyConnectorEntries } from "./connectors/paths.js";
 import {
   parseQmdChunkStrategy,
@@ -352,46 +355,6 @@ function resolveBooleanConfig(
   }
   return coerced;
 }
-function parseGraphPathScoringConfig(
-  cfg: Record<string, unknown> | undefined,
-): {
-  enabled: boolean;
-  invalidNodePenalty: number;
-  includePathInProvenance: boolean;
-} {
-  if (cfg !== undefined && (typeof cfg !== "object" || Array.isArray(cfg))) {
-    throw new Error("graphPathScoring must be a plain object");
-  }
-  const block = cfg ?? {};
-  const invalidNodePenalty = coerceNumber(block.invalidNodePenalty);
-  if (
-    invalidNodePenalty === undefined ||
-    !Number.isFinite(invalidNodePenalty) ||
-    invalidNodePenalty <= 0 ||
-    invalidNodePenalty > 1
-  ) {
-    if (block.invalidNodePenalty !== undefined) {
-      throw new Error(
-        `graphPathScoring.invalidNodePenalty must be a number in (0,1]; got ${JSON.stringify(block.invalidNodePenalty)}`,
-      );
-    }
-  }
-  return {
-    enabled: resolveBooleanConfig(block.enabled, false, "graphPathScoring.enabled"),
-    invalidNodePenalty:
-      invalidNodePenalty !== undefined &&
-      Number.isFinite(invalidNodePenalty) &&
-      invalidNodePenalty > 0 &&
-      invalidNodePenalty <= 1
-        ? invalidNodePenalty
-        : 0.2,
-    includePathInProvenance: resolveBooleanConfig(
-      block.includePathInProvenance,
-      true,
-      "graphPathScoring.includePathInProvenance",
-    ),
-  };
-}
 
 
 function resolvePositiveIntegerConfig(
@@ -548,70 +511,6 @@ function normalizeMemoryRelativeDir(raw: unknown, fallback: string): string {
   return normalized.length > 0 ? normalized : fallback;
 }
 
-/**
- * Parse and validate the semanticChunkingConfig sub-object.
- * Returns only recognized numeric/boolean fields with their correct types.
- */
-function parseContradictionScanConfig(raw: unknown): ContradictionScanConfig {
-  if (!raw || typeof raw !== "object") {
-    return {
-      enabled: false,
-      similarityFloor: 0.82,
-      topicOverlapFloor: 0.4,
-      maxPairsPerRun: 500,
-      cooldownDays: 14,
-      autoMergeDuplicates: false,
-    };
-  }
-  const src = raw as Record<string, unknown>;
-  const simFloor = coerceNumber(src.similarityFloor) ?? 0.82;
-  const topicFloor = coerceNumber(src.topicOverlapFloor) ?? 0.4;
-  const maxPairs = coerceNumber(src.maxPairsPerRun) ?? 500;
-  const cooldown = coerceNumber(src.cooldownDays) ?? 14;
-  return {
-    enabled: coerceBool(src.enabled) === true,
-    similarityFloor: Math.min(1, Math.max(0, simFloor)),
-    topicOverlapFloor: Math.min(1, Math.max(0, topicFloor)),
-    maxPairsPerRun: Math.max(1, maxPairs),
-    cooldownDays: Math.max(0, cooldown),
-    autoMergeDuplicates: coerceBool(src.autoMergeDuplicates) === true,
-  };
-}
-function parseContradictionLocalizationConfig(raw: unknown): ContradictionLocalizationConfig {
-  const defaults: ContradictionLocalizationConfig = {
-    anchorEnabled: true,
-    anchorCandidates: 5,
-    searchCandidates: 5,
-    maxCandidates: 8,
-  };
-  if (raw === undefined || raw === null) return defaults;
-  if (typeof raw !== "object") {
-    throw new Error("contradictionLocalization must be an object");
-  }
-  const src = raw as Record<string, unknown>;
-  const rawEnabled = src.anchorEnabled;
-  const anchorEnabled = rawEnabled === undefined || rawEnabled === null ? true : coerceBool(rawEnabled);
-  if (anchorEnabled === undefined) {
-    throw new Error("contradictionLocalization.anchorEnabled must be a boolean-like value");
-  }
-
-  const parseCap = (key: keyof Pick<ContradictionLocalizationConfig, "anchorCandidates" | "searchCandidates" | "maxCandidates">): number => {
-    const rawValue = src[key];
-    if (rawValue === undefined) return defaults[key];
-    const parsed = coerceNumber(rawValue);
-    if (parsed === undefined || !Number.isInteger(parsed) || parsed < 0) {
-      throw new Error(`contradictionLocalization.${key} must be an integer >= 0`);
-    }
-    return parsed;
-  };
-
-  return {
-    anchorEnabled,
-    anchorCandidates: parseCap("anchorCandidates"),
-    searchCandidates: parseCap("searchCandidates"),
-    maxCandidates: parseCap("maxCandidates"),
-  };
-}
 
 function parseSemanticChunkingConfig(
   raw: unknown,
