+++ packages/remnic-core/src/types.ts
@@ -5,6 +5,10 @@ import type { WearablesConfig } from "./wearables/types.js";
 import type { ExtractionLivenessConfig } from "./extraction-liveness.js";
 import type { ReplicaPeersConfig } from "./replica-peers-config.js";
 import type { ExternalWikiRoot } from "./external-wiki-config.js";
+import type { ContradictionLocalizationConfig, ContradictionScanConfig } from "./contradiction-config.js";
+import type { GraphPathScoringConfig } from "./graph-path-scoring-config.js";
+export type { ContradictionLocalizationConfig, ContradictionScanConfig } from "./contradiction-config.js";
+export type { GraphPathScoringConfig } from "./graph-path-scoring-config.js";
 
 import type { AmbientCaptureProvenance, BufferTurnOwner } from "./buffer-turn-helpers.js";
 export type ReasoningEffort = "none" | "low" | "medium" | "high";
@@ -609,31 +613,6 @@ export interface ScopeTeamConfig {
   promote: string[];
 }
 
-/** Configuration for the nightly contradiction-scan cron (issue #520). */
-export interface ContradictionScanConfig {
-  /** Master switch for the contradiction scan cron. Default true. */
-  enabled: boolean;
-  /** Embedding cosine similarity floor for candidate pair generation. Default 0.82. */
-  similarityFloor: number;
-  /** Minimum topic-token Jaccard overlap for unstructured pairs. Default 0.4. */
-  topicOverlapFloor: number;
-  /** Cap on candidate pairs evaluated per cron run. Default 500. */
-  maxPairsPerRun: number;
-  /** Cooldown in days before re-evaluating a pair judged independent/both-valid. Default 14. */
-  cooldownDays: number;
-  /** When true, pairs judged "duplicates" are auto-flagged for dedup (still need user approval). Default false. */
-  autoMergeDuplicates: boolean;
-}
-export interface ContradictionLocalizationConfig {
-  /** Enable the entity/category anchor pass before QMD search. */
-  anchorEnabled: boolean;
-  /** Maximum anchor candidates. Zero disables the anchor pass. */
-  anchorCandidates: number;
-  /** Maximum QMD search candidates. Zero disables text search. */
-  searchCandidates: number;
-  /** Maximum merged candidates passed to contradiction verification. */
-  maxCandidates: number;
-}
 export interface DependencyPropagationConfig {
   /** Master switch. Default false. */
   enabled: boolean;
@@ -694,11 +673,6 @@ export interface SemanticChunkingConfigShape {
   fallbackToRecursive: boolean;
 }
 
-export interface GraphPathScoringConfig {
-  enabled: boolean;
-  invalidNodePenalty: number;
-  includePathInProvenance: boolean;
-}
 
 export interface PluginConfig extends BoundedJsonlStateConfig {
   openaiApiKey: string | undefined;
